add backend pagination

+++ backend/parking-backend/src/main/java/ru/app/parking_backend/service/ParkingSpaceService.java
@@ -2,22 +2,49 @@ package ru.app.parking_backend.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import ru.app.parking_backend.entity.Client;
+import ru.app.parking_backend.dto.PageResponse;
 import ru.app.parking_backend.entity.ParkingSpace;
 import ru.app.parking_backend.exception.ResourceNotFoundException;
 import ru.app.parking_backend.repository.ParkingSpaceRepository;
+import ru.app.parking_backend.repository.ReservationRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class ParkingSpaceService {
 
     private final ParkingSpaceRepository repository;
+    private final ReservationRepository reservationRepository;
 
-    public List<ParkingSpace> findAll() {
-        return repository.findAll();
+    public PageResponse<ParkingSpace> findAll(String search, int page, int size) {
+        if (page < 0) page = 0;
+        if (size < 10) size = 10;
+        if (size > 100) size = 100;
+
+        int offset = page * size;
+
+        List<ParkingSpace> content;
+        long totalElements;
+
+        if (search != null && !search.trim().isEmpty()) {
+            String searchPattern = search.trim();
+            content = repository.searchByNumberPaginated(searchPattern, size, offset);
+            totalElements = repository.countByNumber(searchPattern);
+        } else {
+            content = repository.findAllPaginated(size, offset);
+            totalElements = repository.countAll();
+        }
+
+        int totalPages = (int) Math.ceil((double) totalElements / size);
+
+        return new PageResponse<>(
+                content,
+                totalElements,
+                totalPages,
+                page,
+                size
+        );
     }
 
     public ParkingSpace findById(Integer id) {
@@ -35,6 +62,9 @@ public class ParkingSpaceService {
     public void delete(Integer id) {
         if (!repository.existById(id)) {
             throw new ResourceNotFoundException("Парковочное место с id " + id + " не найдено");
+        }
+        if (reservationRepository.hasActiveSpace(id)) {
+            throw new IllegalStateException("Невозможно удалить парковочное место: оно в данный момент занято автомобилем");
         }
         repository.delete(id);
     }
